# Optimizations in `symbolic.py`