@geohot 2023-05-25
Approved, but unfortunately I have to reapprove it every time.

Re: optimizations, want to focus on one thing at a time? I see you rewrote a lot of symbolic, are you sure the tests are good enough? There's a lot of subtlety there.