@rayanht 2023-05-25
@geohot I figured in the meantime how to get everything running locally so no worries about the CI.

re: the symbolic stuff, it looks like i changed a lot of things but 90% of it is preserving the exact same logic that was there but moving it around to leverage inheritance instead of calling `isinstance` everywhere because those are expensive af