@geohot 2023-05-25
Yea, smart idea. `isinstance`is terrible.