@rayanht 2023-05-26
# Summary

This diff modifies the `symbolic.py` file to not rely on calls to the expensive `isinstance` by leveraging inheritance to dispatch the correct routines instead.

# Test Plan

## Correctness

Local test suite is passing. Wait for CI

## Performance

`test_net_speed.py` benchmark:

### Before

```
test/test_net_speed.py::TestConvSpeed::test_mnist torch forward pass:  19.002 ms
torch backward pass: 34.680 ms
         108427 function calls (107387 primitive calls) in 58.889 seconds

   Ordered by: internal time
   List reduced from 231 to 46 due to restriction <0.2>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    23100    3.645    0.000    3.645    0.000 {built-in method builtins.isinstance}
      335    2.440    0.007   24.916    0.074 shapetracker.py:185(reshape)
      890    2.139    0.002    3.565    0.004 shapetracker.py:33(__init__)
  460/410    1.676    0.004   41.482    0.101 lazy.py:187(movement_op)
1535/1505    1.475    0.001    3.530    0.002 symbolic.py:28(__mul__)
      685    1.459    0.002   33.434    0.049 shapetracker.py:230(movement_op)
     5405    1.426    0.000    1.426    0.000 shapetracker.py:34(<genexpr>)
      895    1.351    0.002    4.485    0.005 {built-in method builtins.all}
      215    1.276    0.006    8.174    0.038 symbolic.py:109(sum)
      410    1.143    0.003    2.744    0.007 lazy.py:88(__init__)
      955    1.118    0.001    1.916    0.002 symbolic.py:170(create_opnode)
      445    1.068    0.002   49.496    0.111 tensor.py:19(apply)
  665/585    1.046    0.002    2.538    0.004 symbolic.py:43(__floordiv__)
      465    1.041    0.002    4.279    0.009 tensor.py:36(__init__)
      215    1.011    0.005   14.750    0.069 shapetracker.py:61(expr_node)
6065/5705    1.001    0.000    1.124    0.000 {built-in method builtins.len}
      445    0.976    0.002    7.012    0.016 lazy.py:70(create_lazybuffer)
     4280    0.942    0.000    0.942    0.000 shapetracker.py:115(shape)
     2080    0.879    0.000    1.135    0.001 shapetracker.py:187(<genexpr>)
     1925    0.825    0.000    1.174    0.001 enum.py:1229(__hash__)
   420/20    0.766    0.002    2.162    0.108 dataclasses.py:1277(_asdict_inner)
      445    0.728    0.002    1.325    0.003 tensor.py:11(__init__)
      110    0.659    0.006    3.097    0.028 shapetracker.py:214(permute)
     1600    0.645    0.000    0.870    0.001 symbolic.py:118(<lambda>)
      780    0.645    0.001    1.007    0.001 shapetracker.py:215(<genexpr>)
       35    0.640    0.018   17.152    0.490 shapetracker.py:86(merge_views)
      200    0.636    0.003   24.930    0.125 tensor.py:200(reshape)
  425/415    0.564    0.001    1.330    0.003 lazy.py:57(get_weakop)
      585    0.548    0.001    1.263    0.002 symbolic.py:88(__mod__)
     1255    0.515    0.000    0.824    0.001 {built-in method builtins.any}
      825    0.511    0.001    0.663    0.001 shapetracker.py:107(__init__)
     1045    0.500    0.000    0.611    0.001 shapetracker.py:180(<genexpr>)
      380    0.491    0.001    1.178    0.003 weakref.py:164(__setitem__)
     1300    0.475    0.000    0.616    0.000 shapetracker.py:194(<genexpr>)
      970    0.465    0.000    0.647    0.001 shapetracker.py:112(contiguous)
  565/545    0.449    0.001    0.746    0.001 symbolic.py:13(render)
       20    0.437    0.022   24.347    1.217 tensor.py:331(_pool)
      430    0.433    0.001    1.028    0.002 helpers.py:15(<listcomp>)
      955    0.428    0.000    0.546    0.001 symbolic.py:156(create_node)
      430    0.418    0.001    2.474    0.006 helpers.py:15(partition)
      320    0.416    0.001    0.639    0.002 copy.py:128(deepcopy)
  920/890    0.413    0.000    0.518    0.001 lazy.py:57(<genexpr>)
      870    0.405    0.000    0.586    0.001 <string>:1(<lambda>)
      175    0.403    0.002    2.019    0.012 shapetracker.py:179(expand)
  420/410    0.396    0.001    0.847    0.002 ops.py:28(get_buffers)
      215    0.391    0.002    2.438    0.011 symbolic.py:126(<listcomp>)


forward pass:  11.930 ms, 0.63x off baseline 19.002 ms
backward pass: 0.141 ms, 0.00x off baseline 34.680 ms
```

### After

```
test/test_net_speed.py::TestConvSpeed::test_mnist torch forward pass:  20.084 ms
torch backward pass: 23.449 ms
         93792 function calls (92862 primitive calls) in 53.364 seconds

   Ordered by: internal time
   List reduced from 225 to 45 due to restriction <0.2>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      335    2.460    0.007   18.745    0.056 shapetracker.py:186(reshape)
      890    2.128    0.002    3.540    0.004 shapetracker.py:34(__init__)
    12045    1.942    0.000    1.942    0.000 {built-in method builtins.isinstance}
      215    1.889    0.009    5.075    0.024 symbolic.py:59(sum)
  460/410    1.654    0.004   35.413    0.086 lazy.py:187(movement_op)
      685    1.533    0.002   27.355    0.040 shapetracker.py:231(movement_op)
     5405    1.412    0.000    1.412    0.000 shapetracker.py:35(<genexpr>)
      895    1.130    0.001    4.009    0.004 {built-in method builtins.all}
      410    1.119    0.003    2.746    0.007 lazy.py:88(__init__)
      445    1.085    0.002   43.460    0.098 tensor.py:19(apply)
      465    1.071    0.002    4.323    0.009 tensor.py:36(__init__)
6065/5705    1.056    0.000    1.186    0.000 {built-in method builtins.len}
     4280    1.004    0.000    1.004    0.000 shapetracker.py:116(shape)
      215    0.991    0.005    9.283    0.043 shapetracker.py:62(expr_node)
     1650    0.987    0.001    2.344    0.001 symbolic.py:30(__mul__)
      445    0.966    0.002    7.022    0.016 lazy.py:70(create_lazybuffer)
     1925    0.848    0.000    1.176    0.001 enum.py:1229(__hash__)
     1345    0.815    0.001    1.365    0.001 symbolic.py:120(__init__)
   420/20    0.759    0.002    2.186    0.109 dataclasses.py:1277(_asdict_inner)
      445    0.751    0.002    1.329    0.003 tensor.py:11(__init__)
       35    0.689    0.020   11.409    0.326 shapetracker.py:87(merge_views)
      200    0.675    0.003   20.623    0.103 tensor.py:200(reshape)
      780    0.674    0.001    1.048    0.001 shapetracker.py:216(<genexpr>)
      110    0.658    0.006    3.097    0.028 shapetracker.py:215(permute)
     1430    0.612    0.000    0.774    0.001 shapetracker.py:188(<genexpr>)
  670/650    0.599    0.001    0.910    0.001 symbolic.py:15(render)
  425/415    0.564    0.001    1.339    0.003 lazy.py:57(get_weakop)
      825    0.538    0.001    0.672    0.001 shapetracker.py:108(__init__)
     1045    0.530    0.001    0.660    0.001 shapetracker.py:181(<genexpr>)
      380    0.505    0.001    1.144    0.003 weakref.py:164(__setitem__)
      970    0.504    0.001    0.674    0.001 shapetracker.py:113(contiguous)
     1040    0.487    0.000    0.785    0.001 {built-in method builtins.any}
     1300    0.454    0.000    0.592    0.000 shapetracker.py:195(<genexpr>)
     1125    0.449    0.000    0.449    0.000 symbolic.py:145(get_bounds)
      320    0.441    0.001    0.647    0.002 copy.py:128(deepcopy)
     1115    0.434    0.000    0.550    0.000 symbolic.py:114(create_node)
       20    0.423    0.021   20.026    1.001 tensor.py:331(_pool)
  420/410    0.420    0.001    0.871    0.002 ops.py:28(get_buffers)
      175    0.411    0.002    2.016    0.012 shapetracker.py:180(expand)
  920/890    0.409    0.000    0.521    0.001 lazy.py:57(<genexpr>)
      870    0.407    0.000    0.599    0.001 <string>:1(<lambda>)
     1560    0.397    0.000    0.397    0.000 shapetracker.py:193(<genexpr>)
     1310    0.393    0.000    0.393    0.000 symbolic.py:111(__init__)
     2460    0.388    0.000    0.388    0.000 {method 'append' of 'list' objects}
       60    0.376    0.006    4.316    0.072 tensor.py:213(slice)


forward pass:  10.745 ms, 0.54x off baseline 20.084 ms
backward pass: 0.142 ms, 0.01x off baseline 23.449 ms
```

We roughly halve the number of calls to `isinstance` from 23100 to 12045. Performance gain on this benchmark is on average 10-20%